+++ src/main/java/com/wztwzt/ae2_qof/network/ModNetwork.java
@@ -87,5 +87,9 @@ public final class ModNetwork {
             SmartDoublingTogglePacket.class,
             discriminator++,
             Side.SERVER);
+
+        CHANNEL.registerMessage(ConfigSetPacket.Handler.class, ConfigSetPacket.class, discriminator++, Side.SERVER);
+
+        CHANNEL.registerMessage(ConfigUpdatePacket.Handler.class, ConfigUpdatePacket.class, discriminator++, Side.CLIENT);
     }
 }